+++ MixxFit.API/Features/WeightEntries/GetWeightSummary/CurrentWeightDto.cs
@@ -2,6 +2,6 @@ namespace MixxFit.API.Features.WeightEntries.GetWeightSummary;
 
 public record CurrentWeightDto
 {
-    public double? Weight { get; set; }
+    public decimal? Weight { get; set; }
     public DateTime? CreatedAt { get; set; }
 }
