+++ MixxFit.API/Features/WeightEntries/GetWeightSummary/WeightDeltaDto.cs
@@ -2,6 +2,6 @@ namespace MixxFit.API.Features.WeightEntries.GetWeightSummary;
 
 public record WeightDeltaDto
 {
-    public double Delta { get; set; }
+    public decimal Delta { get; set; }
     public DateTime CreatedAt { get; set; }
 };
